Implemented MAGETWO-16634: Merge story branch into the teams develop branch

+++ dev/tests/static/testsuite/Magento/Test/Integrity/Di/CompilerTest.php
@@ -78,7 +78,9 @@ class CompilerTest extends \PHPUnit_Framework_TestCase
     protected function tearDown()
     {
         $filesystem = new \Magento\Filesystem\Driver\File();
-        $filesystem->deleteDirectory($this->_tmpDir);
+        if ($filesystem->isExists($this->_tmpDir)) {
+            $filesystem->deleteDirectory($this->_tmpDir);
+        }
     }
 
     /**
